requirements.txt
pandas 
psycopg2-binary 
jupyter 
python-dotenv
kaggle
pyarrow